Refactor code into ui and main packages
