+++ src/main/java/main/Storage.java
@@ -1,4 +1,4 @@
-package main.java;
+package main;
 
 import java.io.File;
 import java.io.IOException;
