+++ src/test/java/UiTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import main.java.Ui;
+import ui.Ui;
 
 public class UiTest {
